requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pdfplumber==0.11.4
pypdf2==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==43.0.3
apscheduler==3.10.4
httpx==0.27.2
yfinance==0.2.50
openpyxl==3.1.5
scipy==1.14.1
python-multipart==0.0.12
pydantic[email]==2.9.2
python-dotenv==1.0.1
jinja2==3.1.4
qrcode[pil]==8.0
Pillow==11.0.0
netifaces==0.11.0
